Tests/LibWeb/Text/expected/css/animation-timing-function-var-in-keyframe.txt

var(linear) at 0.5: 100px
var(undefined, linear) at 0.5: 100px
var(undefined) at 0.5: 100px
